# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Kanban REST API
  description: |
    Comprehensive REST API for the MCP Kanban Task Management System.
    
    This API provides full CRUD operations for boards, tasks, notes, tags, and more.
    It supports real-time updates via WebSocket, AI-powered context generation,
    and advanced task management features.
    
    ## Features
    - **Task Management**: Create, update, delete, and organize tasks
    - **Subtask Support**: Hierarchical task organization with progress tracking
    - **Task Dependencies**: Dependency management with cycle detection
    - **Task Templates**: Predefined templates for common task types
    - **Board Management**: Multi-board support with customizable columns
    - **Real-time Updates**: WebSocket integration for live collaboration
    - **AI Integration**: Context-aware task prioritization and suggestions
    - **Analytics & Reporting**: Completion analytics and velocity tracking
    - **Performance Monitoring**: System health and performance dashboards
    - **Backup & Recovery**: Comprehensive data backup and restoration
    - **Advanced Search**: Full-text search across tasks and notes
    - **Tag System**: Hierarchical tagging with automatic categorization
    - **Git Integration**: Branch-based board mapping and context awareness
    
    ## Authentication
    All API requests require authentication using API keys.
    
    ## Rate Limiting
...
    
    ## WebSocket Support
...
    
    ## MCP Integration
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /boards:
    get:
      summary: List Boards
      description: Retrieve a list of boards with optional filtering and pagination
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: name
          in: query
          description: Filter boards by name (partial match)
          schema:
            type: string
        - name: status
          in: query
          description: Filter boards by status
          schema:
            type: string
            enum: [active, archived]
      responses:
        '200':
          description: List of boards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create Board
      description: Create a new board with specified columns and settings
      tags: [Boards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '201':
          description: Board created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /boards/{boardId}:
    get:
      summary: Get Board
      description: Retrieve a specific board with its tasks and columns
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
        - name: includeTasks
          in: query
          description: Include tasks in the response
          schema:
            type: boolean
            default: true
        - name: includeArchived
          in: query
          description: Include archived tasks
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Board details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Update Board
      description: Update board properties and settings
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
      responses:
        '200':
          description: Board updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete Board
      description: Delete a board and optionally its tasks
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
        - name: cascade
          in: query
          description: Delete all tasks in the board
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Board deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /boards/{boardId}/columns:
    get:
      summary: List Board Columns
      description: Get all columns for a specific board
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
      responses:
        '200':
          description: List of columns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create Board Column
      description: Add a new column to a board
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
      responses:
        '201':
          description: Column created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /boards/{boardId}/columns/{columnId}:
    patch:
      summary: Update Board Column
      description: Update column properties
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
        - $ref: '#/components/parameters/ColumnIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateColumnRequest'
      responses:
        '200':
          description: Column updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete Board Column
      description: Remove a column from a board
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
        - $ref: '#/components/parameters/ColumnIdParam'
        - name: moveTasksTo
          in: query
          description: Column ID to move tasks to
          schema:
            type: string
      responses:
        '204':
          description: Column deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /boards/{boardId}/archive:
    post:
      summary: Archive Board
      description: Archive a board and optionally its tasks
      tags: [Boards]
      parameters:
        - $ref: '#/components/parameters/BoardIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveBoardRequest'
      responses:
        '200':
          description: Board archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key in Bearer format (e.g., "Bearer your-api-key")
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [created_at, updated_at, name, title, priority, due_date]
        default: created_at
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    BoardIdParam:
      name: boardId
      in: path
      required: true
      description: Board identifier
      schema:
        type: string
        pattern: '^board-[a-zA-Z0-9-]+$'
...
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - resource already exists or constraint violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateBoardRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        columns:
          type: array
          items:
            $ref: '#/components/schemas/CreateColumnRequest'
          minItems: 1
          maxItems: 10
      required:
        - name
    UpdateBoardRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [active, archived]
    BoardResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Board'
    BoardListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Board'
        pagination:
          $ref: '#/components/schemas/Pagination'
...